ArgsTable: Refactor no controls warning & add stories
