+++ addons/controls/src/components/ControlsPanel.tsx
@@ -1,33 +1,13 @@
 import React, { FC } from 'react';
-import { styled } from '@storybook/theming';
-import { ArgsTable, Link } from '@storybook/components';
+import { ArgsTable, NoControlsWarning } from '@storybook/components';
 import { useArgs, useArgTypes, useParameter } from '@storybook/api';
+
 import { PARAM_KEY } from '../constants';
 
 interface ControlsParameters {
   expanded?: boolean;
   hideNoControlsWarning?: boolean;
 }
-
-const NoControlsWrapper = styled.div(({ theme }) => ({
-  background: theme.background.warning,
-  padding: '10px 15px',
-  lineHeight: '20px',
-  boxShadow: `${theme.appBorderColor} 0 -1px 0 0 inset`,
-}));
-
-const NoControlsWarning = () => (
-  <NoControlsWrapper>
-    This story is not configured to handle controls.&nbsp;
-    <Link
-      href="https://github.com/storybookjs/storybook/blob/next/addons/controls/README.md#writing-stories"
-      target="_blank"
-      cancel={false}
-    >
-      Learn how to add controls »
-    </Link>
-  </NoControlsWrapper>
-);
 
 export const ControlsPanel: FC = () => {
   const [args, updateArgs] = useArgs();
